+++ src/FluentAssertions.Analyzers/Constants.cs
@@ -1,17 +1,10 @@
-﻿namespace FluentAssertions.Analyzers
+namespace FluentAssertions.Analyzers
 {
     public static class Constants
     {
         public static class DiagnosticProperties
         {
             public const string Title = nameof(Title);
-            public const string VariableName = nameof(VariableName);
-            public const string LambdaString = nameof(LambdaString);
-            public const string ExpectedItemString = nameof(ExpectedItemString);
-            public const string UnexpectedItemString = nameof(UnexpectedItemString);
-            public const string BecauseArgumentsString = nameof(BecauseArgumentsString);
-            public const string ArgumentString = nameof(ArgumentString);
-
             public const string VisitorName = nameof(VisitorName);
         }
