+++ src/FluentAssertions.Analyzers/Tips/Collections/CollectionShouldNotContainItem.cs
@@ -40,10 +40,10 @@ namespace FluentAssertions.Analyzers
         
         protected override ExpressionSyntax GetNewExpression(ExpressionSyntax expression, FluentAssertionsDiagnosticProperties properties)
         {
-            var remove = new NodeReplacement.RemoveAndExtractArgumentsNodeReplacement("Contains");
+            var remove = NodeReplacement.RemoveAndExtractArguments("Contains");
             var newExpression = GetNewExpression(expression, remove);
 
-            return GetNewExpression(newExpression, new NodeReplacement.RenameAndPrependArgumentsNodeReplacement("BeFalse", "NotContain", remove.Arguments));
+            return GetNewExpression(newExpression, NodeReplacement.RenameAndPrependArguments("BeFalse", "NotContain", remove.Arguments));
         }
     }
 }
