+++ src/FluentAssertions.Analyzers/Utilities/FluentAssertionsAnalyzer.cs
@@ -88,6 +88,5 @@ namespace FluentAssertions.Analyzers
 
     public abstract class FluentAssertionsAnalyzer : FluentAssertionsAnalyzer<FluentAssertionsCSharpSyntaxVisitor>
     {
-        protected override IEnumerable<FluentAssertionsCSharpSyntaxVisitor> Visitors => Enumerable.Empty<FluentAssertionsCSharpSyntaxVisitor>();
     }
 }
