+++ src/FluentAssertions.Analyzers/Utilities/MemberValidator.cs
